refactor(worker): define folder structure
