+++ worker/requirements.txt
@@ -7,3 +7,4 @@ click==8.1.7
 ffmpeg-python==0.2.0
 pydantic==2.10.2
 requests==2.32.3
+pydantic-settings==2.7.1
