--- worker/run_ml_tests.sh
@@ -1,1 +1,0 @@
-docker compose run --build --remove-orphans -it worker python -m pytest -m ml -s
