+++ worker/worker/utils/api.py
@@ -2,18 +2,18 @@ import typing as tp
 
 import requests
 
-from .settings import worker_config
-from .enums import TranscriptionState
-from .schemas import CreateTranscriptionChunk
+from worker.core.settings import settings
+from worker.core.enums import TranscriptionState
+from worker.core.schemas import CreateTranscriptionChunk
 
 
 def get_transcription_state_chunk_size_secs(transcription_id: int) -> TranscriptionState:
-    url = f"{worker_config.API_BASE_URL}/worker/transcription_status"
+    url = f"{settings.api_base_url}/worker/transcription_status"
     headers = {
         "accept": "application/json",
         "Content-Type": "application/json",
     }
-    params = {"secret_worker_token": worker_config.SECRET_WORKER_TOKEN}
+    params = {"secret_worker_token": settings.secret_worker_token}
     json_data = {"transcription_id": transcription_id}
     print("[API CALL] Transcription State Update (to get status)...")
     response = requests.post(
@@ -32,12 +32,12 @@ def update_transcription_state(
     transcription_state: tp.Optional[TranscriptionState] = None,
     transcription_chunk: tp.Optional[CreateTranscriptionChunk] = None,
 ) -> str:
-    url = f"{worker_config.API_BASE_URL}/worker/transcription_status"
+    url = f"{settings.api_base_url}/worker/transcription_status"
     headers = {
         "accept": "application/json",
         "Content-Type": "application/json",
     }
-    params = {"secret_worker_token": worker_config.SECRET_WORKER_TOKEN}
+    params = {"secret_worker_token": settings.secret_worker_token}
     json_data = {
         "transcription_id": transcription_id,
     }
